Create alpha_strategy.h
This header file declares an abstract class called AlphaStrategy, which defines an interface for a trading strategy that captures alpha. The class defines two methods:
interval(): This method returns the interval between updates in milliseconds. The bot will use this interval to sleep between iterations.
getSignals(): This method returns the trading signals generated by the strategy. The signals will be used by the bot to place orders.