Create signal.h
This header file declares a struct called Signal which is used to store the type of the signal, the pair, the price and the size of the order. The type of the signal can be either BUY or SELL. The pair is a string that represents the crypto currency pair, such as BTC/USD. The price is a double, representing the price at which the order should be placed. The size is an int, representing the amount of the crypto currency to be bought or sold.