Create exchange_api.h
placeOrder method uses the libcurl library to send a POST request to the exchange's API to place an order. It sets the pair, price, size, and order type in the request body and sets the API key and secret key for authentication. If the request is successful, it will place the order in the exchange, otherwise it will print an error message to the console.